+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/claims/modifiers/RegexReplaceModifier.java
@@ -9,9 +9,11 @@ import java.util.regex.Pattern;
 
 /**
  * Replace regex modifier. Replaces parts matched by regex with string using backreferences to groups.
+ *
+ * Configuration (replace [claimName] with the name of the claim and [modifierName] with the name of modifier):
  * <ul>
- *     <li><b>custom.claim.[claimName].modifier.find</b> - string to be replaced, can be a regex</li>
- *     <li><b>custom.claim.[claimName].modifier.append</b> - string to be used as replacement</li>
+ *     <li><b>custom.claim.[claimName].modifier.[modifierName].find</b> - string to be replaced, can be a regex</li>
+ *     <li><b>custom.claim.[claimName].modifier.[modifierName].replace</b> - string to be used as replacement</li>
  * </ul>
  *
  * @see java.util.regex.Matcher#replaceAll(String)
@@ -27,27 +29,26 @@ public class RegexReplaceModifier extends ClaimModifier {
 
 	private final Pattern regex;
 	private final String replacement;
-	private final String claimName;
 
 	public RegexReplaceModifier(ClaimModifierInitContext ctx) {
 		super(ctx);
-		this.claimName = ctx.getClaimName();
 		regex = Pattern.compile(ctx.getProperty(FIND, ""));
 		replacement = ctx.getProperty(REPLACE, "");
-		log.debug("{}(modifier) - regex: '{}', replacement: '{}'", claimName, regex, replacement);
+		log.debug("{}(modifier) - regex: '{}', replacement: '{}'", getUnifiedName(), regex, replacement);
 	}
 
 	@Override
 	public String modify(String value) {
 		String modified = regex.matcher(value).replaceAll(replacement);
-		log.trace("{} - modifying value '{}' by replacing matched part ('{}') with: '{}'", claimName, value, regex,
-				replacement);
-		log.trace("{} - new value: '{}", claimName, modified);
+		log.trace("{} - modifying value '{}' by replacing matched part ('{}') with: '{}'", getUnifiedName(),
+				value, regex, replacement);
+		log.trace("{} - new value: '{}", getUnifiedName(), modified);
 		return modified;
 	}
 
 	@Override
 	public String toString() {
-		return "RegexReplaceModifier replacing" + regex.pattern() + " with " + replacement;
+		return getUnifiedName() + " - RegexReplaceModifier replacing '" + regex.pattern()
+				+ "' with '" + replacement + '\'';
 	}
 }