+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/claims/sources/EdupersonScopedAffiliationsMUSource.java
@@ -13,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -48,17 +49,49 @@ public class EdupersonScopedAffiliationsMUSource extends ClaimSource {
 	private final Map<List<Long>, String> affiliations = new HashMap<>();
 	private Long voId = 363L;
 	private String valueScope = "muni.cz";
-	private final String claimName;
 
 	public EdupersonScopedAffiliationsMUSource(ClaimSourceInitContext ctx) {
 		super(ctx);
-		this.claimName = ctx.getClaimName();
 		parseConfigFile(ctx.getProperty(CONFIG_FILE, DEFAULT_PATH));
-		log.debug("{} - affiliations: '{}', voId: '{}', valueScope: '{}'", claimName, affiliations, voId, valueScope);
+		log.debug("{} - affiliations: '{}', voId: '{}', valueScope: '{}'",
+				getClaimName(), affiliations, voId, valueScope);
+	}
+
+	@Override
+	public Set<String> getAttrIdentifiers() {
+		return Collections.emptySet();
+	}
+
+	@Override
+	public JsonNode produceValue(ClaimSourceProduceContext pctx) {
+		Long userId = pctx.getPerunUserId();
+		ArrayNode result = JsonNodeFactory.instance.arrayNode();
+		Set<Long> groups = pctx.getPerunAdapter().getUserGroupsIds(userId, voId);
+		for (Map.Entry<List<Long>, String> entry : affiliations.entrySet()) {
+			for (Long id: entry.getKey()) {
+				if (groups.contains(id)) {
+					String affiliation = entry.getValue() + '@' + valueScope;
+					log.trace("{} - added affiliation '{}' due to membership in group '{}'",
+							getClaimName(), affiliation, id);
+					result.add(affiliation);
+					break;
+				}
+			}
+		}
+
+		if (result.size() == 0) {
+			String affiliation = "affiliate@" + valueScope;
+			log.trace("{} - user is not a member in any special groups, added default affiliation: '{}'",
+					getClaimName(), affiliation);
+			result.add(affiliation);
+		}
+
+		log.debug("{} - produced value for user({}): '{}'", getClaimName(), userId, result);
+		return result;
 	}
 
 	private void parseConfigFile(String file) {
-		log.trace("{} - Loading config file {}", claimName, file);
+		log.trace("{} - Loading config file {}", getClaimName(), file);
 		YAMLMapper mapper = new YAMLMapper();
 		try {
 			JsonNode root = mapper.readValue(new File(file), JsonNode.class);
@@ -73,34 +106,8 @@ public class EdupersonScopedAffiliationsMUSource extends ClaimSource {
 				affiliations.put(gids, value);
 			}
 		} catch (IOException ex) {
-			log.warn("{} - cannot read claim configuration file: '{}'", claimName, file);
+			log.warn("{} - cannot read claim configuration file: '{}'", getClaimName(), file);
 		}
 	}
 
-	@Override
-	public JsonNode produceValue(ClaimSourceProduceContext pctx) {
-		Long userId = pctx.getPerunUserId();
-		ArrayNode result = JsonNodeFactory.instance.arrayNode();
-		Set<Long> groups = pctx.getPerunAdapter().getUserGroupsIds(userId, voId);
-		for (Map.Entry<List<Long>, String> entry : affiliations.entrySet()) {
-			for (Long id: entry.getKey()) {
-				if (groups.contains(id)) {
-					String affiliation = entry.getValue() + '@' + valueScope;
-					log.trace("{} - added affiliation '{}' due to membership in group '{}'", claimName, affiliation, id);
-					result.add(affiliation);
-					break;
-				}
-			}
-		}
-
-		if (result.size() == 0) {
-			String affiliation = "affiliate@" + valueScope;
-			log.trace("{} - user is not a member in any special groups, added default affiliation: '{}'",
-					claimName, affiliation);
-			result.add(affiliation);
-		}
-
-		log.debug("{} - produced value for user({}): '{}'", claimName, userId, result);
-		return result;
-	}
 }